Cleanup and restructure
- Remove unused files `_config.yml`, `run.sh`, `logo.xcf`
- Extract static and githubWebHook routes with blueprints from `server.py`
- Remove old restart logic
- Improve tests
