--- run.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash
-
-git pull
-. .env
-pip install -r requirements.txt
-python src/server.py
